Developer Manual

This page is a big TODO!

For now, there are only the doc-strings:

Morbit.ExactConfig — Type

ExactConfig(; gradients, jacobian = nothing, max_evals = typemax(Int64))

Configuration for an ExactModel. gradients should be a vector of callbacks for the objective gradients or a Symbol, either :autodiff or fdm, to define the differentiation method to use on the objective. Alternatively, a jacobian handle can be provided.

Morbit.ExactModel — Type

ExactModel( tfn, objf, diff_fn )

Exact Model type for evaluating the objective function objf directly. Is instantiated by the corresponding init_model and update_model functions.

Morbit.RbfConfig — Type

RbfConfig(; kwarg1 = val1, … )

Configuration type for local RBF surrogate models.

To choose a kernel, use the kwarg kernel and a value of either :cubic (default), :inv_multiquadric, :multiquadric, :gaussian or :thin_plate_spline. The kwarg shape_parameter takes a constant number or a string that defines a calculation on Δ, e.g, "Δ/10". Note, that shape_parameter has a different meaning for the different kernels. For `:gaussian, :inv_multiquadric, :multiquadric it actually is a floating point shapeparameter. For :cubic it is the (odd) integer exponent and for `thinplate_splineit is an integer exponent as well. UseNaN` for defaults.

To see other configuration parameters use fieldnames(Morbit.RbfConfig). They have individual docstrings attached.

Morbit.TaylorApproximateConfig — Type

TaylorApproximateConfig(;degree=1,mode=:fdm,max_evals=typemax(Int64))

Configure a linear or quadratic Taylor model where the gradients and Hessians are constructed either by finite differencing (mode = :fdm) or automatic differencing (mode = :autodiff).

Morbit.TaylorCallbackConfig — Type

TaylorCallbackConfig(;degree=1,gradients,hessians=nothing,max_evals=typemax(Int64))

Configuration for a linear or quadratic Taylor model where there are callbacks provided for the gradients and – if applicable – the Hessians. The gradients keyword point to an array of callbacks where each callback evaluates the gradient of one of the outputs.

Morbit.TaylorConfig — Type

TaylorConfig(; degree, gradients :: RFD.CFDStamp, hessians :: RFD.CFDStamp, max_evals)

Configuration for a polynomial Taylor model using finite difference approximations of the derivatives. By default we have degree = 2 and gradients == hessians == RFD.CFDStamp(1,2), that is, a first order central difference scheme of accuracy order 3 is recursed to compute the Hessians and the gradients. In this case, the finite difference scheme is the same for both Hessians and gradients and we profit from caching intermediate results.

Morbit.TransformerFn — Type

Return the TransformerFn defined by mop with a minimum precision of T.

Morbit.TransformerFn — Method

Unscale the point x̂ from internal to original domain.

Base.eachindex — Method

List of all id :: Int belonging to the stored results.

Base.length — Method

Number of entries in database.

Morbit._add! — Method

Add an objective function to MOP with specified output indices.

Morbit._backtrack — Method

Perform a backtracking loop starting at x with an initial step of step_size .* dir and return trial point x₊, the surrogate value-vector m_x₊ and the final step s = x₊ .- x.

Morbit._consume_points — Method

_consume_points(data_base, poised_points, poised_indices, candidate_indices)

Helper to return array of database indices for poised_points and poised_indices. Add result to database if index is -1. candidate_indices are the database indices of the points from the trust region.

Morbit._del! — Method

Remove an objective function from MOP.

Morbit._eval_models — Method

Evaluate (internal) output ℓ of TaylorModel tm, provided a difference vector h = x - x0.

Morbit._get_optim_handle — Method

Return a function handle to be used with NLopt for output ℓ of model. That is, if model is a surrogate for two scalar objectives, then ℓ must be either 1 or 2.

Morbit._get_ps_constraint_func — Method

_get_ps_constraint_func( sc :: SurrogateContainer, mx, dir, l )

Return the l-th (possibly non-linear) constraint function for Pascoletti-Serafini. dir .>= 0 is the image direction; χ = [t;x] is the augmented variable vector;

Morbit._get_ps_objective_func — Method

Return objective function for Pascoletti-Serafini, modifying gradient in place.

Morbit._init_model — Method

Return an ExactModel build from a VectorObjectiveFunction objf. Model is the same inside and outside of criticality round.

Morbit._intersect_bounds — Method

Return smallest positive and biggest negative and σ₊ and σ₋ so that x .+ σ± .* d stays within bounds.

Morbit._local_bounds — Method

Return lower and upper bound vectors combining global and trust region constraints.

Morbit._objf_index — Method

Position of objf in list_of_objectives(mop).

Morbit._ps_optimization — Method

Construct and solve Pascoletti Serafini subproblem using surrogates from sc.

Morbit._steepest_descent_direction — Method

Provided x and the (surrogate) jacobian ∇F at x, as well as bounds lb and ub, return steepest multi descent direction.

Morbit._wrap_func — Method

A general constructor.

Morbit._wrap_func — Method

A general constructor.

Morbit.add_objective! — Function

add_objective!( mop :: MixedMOP, func :: T where{T <: Function}, type :: Symbol = :expensive, n_out :: Int64 = 1, can_batch :: Bool = false )

Add scalar-valued objective function func to mop structure. func must take an Vec as its (first) argument, i.e. represent a function $f: ℝ^n → ℝ$. type must either be :expensive or :cheap to determine whether the function is replaced by a surrogate model or not.

If type is :cheap and func takes 1 argument only then its gradient is calculated by ForwardDiff. A cheap function func with custom gradient function grad (representing $∇f : ℝ^n → ℝ^n$) is added by

add_objective!(mop, func, grad)

The optional argument n_out allows for the specification of vector-valued objective functions. This is mainly meant to be used for expensive functions that are in some sense inter-dependent.

The flag can_batch defaults to false so that the objective function is simply looped over a bunch of arguments if required. If can_batch == true then the objective function must be able to return an array of results when provided an array of input vectors (whilst still returning a single result, not a singleton array containing the result, for a single input vector).

Examples

# Define 2 scalar objective functions and a MOP ℝ^2 → ℝ^2

f1(x) =  x[1]^2 + x[2]

f2(x) = exp(sum(x))
∇f2(x) = exp(sum(x)) .* ones(2);

mop = MixedMOP()
add_objective!(mop, f1, :cheap)     # gradient will be calculated using ForwardDiff
add_objective!(mop, f2, ∇f2 )       # gradient is provided

Morbit.add_objective! — Method

add_objective!( mop :: MixedMOP, func :: T where{T <: Function}, grad :: T where{T <: Function})

Add scalar-valued objective function func and its vector-valued gradient grad to mop struture.

Morbit.add_objective! — Method

Add a scalar objective to mop::MixedMOP modelled according to model_config.

Morbit.add_vector_objective! — Method

Add a vector objective to mop::MixedMOP modelled according to model_config.

Morbit.combine — Method

Combine two objectives. Only needed if combinable can return true.

Morbit.combine — Method

Get a new function function handle stacking the output of func1 and func2.

Morbit.copy_db — Method

Return a new database of same 'base' type but with different saveable type.

Morbit.eval_all_objectives — Method

(Internally) Evaluate all objectives at site x̂::Vec. Objective order might differ from order in which they were added.

Morbit.eval_and_sort_objectives — Method

Evaluate all objectives at site x̂::Vec and sort the result according to the order in which objectives were added.

Morbit.eval_handle — Method

Return a function that evaluates an objective at an unscaled site.

Morbit.eval_models — Method

Evaluate output ℓ of the ExactModel em at scaled site x̂.

Morbit.eval_models — Method

Evaluate the ExactModel em at scaled site x̂.

Morbit.eval_models — Method

Return model value for output l of sc at x̂. Index l is assumed to be an internal index in the range of 1,…,nobjfs, where nobjfs is the total number of (scalarized) objectives stored in sc.

Morbit.eval_models — Method

Evaluate (internal) output ℓ of tm at scaled site x̂.

Morbit.eval_models — Method

Evaluate output ℓ of mod::RbfModel at scaled site x̂.

Morbit.eval_models — Method

Evaluate mod::RbfModel at scaled site x̂.

Morbit.eval_objf — Method

Evaluate the objective at unscaled site x. and increase counter.

Morbit.eval_objf — Method

Evaluate the objective at scaled site x̂ with help of tfn. Used in diff_wrappers.

Morbit.full_lower_bounds — Method

Return full vector of lower variable vectors for original problem.

Morbit.full_lower_bounds_internal — Method

Return lower variable bounds for scaled variables.

Morbit.full_upper_bounds — Method

Return full vector of upper variable vectors for original problem.

Morbit.full_upper_bounds_internal — Method

Return upper variable bounds for scaled variables.

Morbit.get_gradient — Method

Gradient vector of output ℓ of em at scaled site x̂.

Morbit.get_gradient — Method

Return a gradient for output l of sc at x̂. Index l is assumed to be an internal index in the range of 1,…,nobjfs, where nobjfs is the total number of (scalarized) objectives stored in sc.

Morbit.get_gradient — Method

Gradient vector of output ℓ of mod at scaled site x̂.

Morbit.get_id — Method

get_id( res :: AbstractResult ) :: Int

Return the id of a result such that for the database db conataining res it holds that get_result(db, id) == res.

Morbit.get_jacobian — Method

Jacobian Matrix of ExactModel em at scaled site x̂.

Morbit.get_jacobian — Method

Jacobian Matrix of ExactModel em at scaled site x̂.

Morbit.get_optim_handle — Method

Return a function handle to be used with NLopt for output l of sc. Index l is assumed to be an internal index in the range of 1, …, n_objfs, where n_objfs is the total number of (scalarized) objectives stored in sc.

Morbit.get_poised_set — Method

get_poised_set( basis, points; solver = :LN_BOBYQA, max_solver_evals = -1 )

Compute a point set suited for polynomial interpolation.

Input:

basis: A vector of polynomials constituting a basis for the polynomial space.

points: (optional) A set of candidate points to be tried for inclusion into the poised set.

solver: NLopt solver to use. Should be derivative-free.

max_solver_evals: Maximum number of evaluations in each optimization run.

Return:

poised_points :: Vector{T} where T is either a Vector{F} or an SVector{n_vars, F} and F is the precision of the points in points, but at least Float32.

lagrange_basis :: Vector{<:AbstractPolynomialLike}: The Lagrange basis corresponding to poised_points.

point_indices: An array indicating which points from points are also in poised_points. A positive entry corresponds to the index of a poised point in points. If a poised point is new, then the entry is -1.

Morbit.get_result — Method

Get result with id from database db.

Morbit.get_scaling_poly — Method

Return vector of polynomials that scales variables from [lb, ub] to [0,1]^n.

Morbit.get_site — Method

Return the site vector associated with a result.

Morbit.get_unscaling_poly — Method

Return vector of polynomials that unscales variables from [0,1]^n to [lb,ub].

Morbit.get_value — Method

Return the value vector associated with a result.

Morbit.grow_radius — Method

Grow radius according to min( Δ_max, γ * Δ ).

Morbit.grow_radius — Method

Grow radius according to min( Δ_max, (γ + ||s||/Δ) * Δ )

Morbit.inc_evals! — Function

Increase evaluation count by N

Morbit.init_db — Method

Constructor for empty database of type T.

Morbit.init_res — Method

Constructor for a result, taking site and value vector and id in database.

Morbit.init_surrogates — Method

Return a SurrogateContainer initialized from the information provided in mop.

Morbit.is_transformed — Method

Bool indicating if the database data been transformed.

Morbit.list_of_objectives — Method

Return a list of AbstractVectorObjectives.

Morbit.local_bounds — Method

Local bounds vectors lb_eff and ub_eff using scaled variable constraints from mop.

Morbit.make_set_lambda_poised — Method

make_set_lambda_poised( basis, points; 
    LAMBDA = 1.5, solver = :LN_BOBYQA, max_solver_evals = -1, max_loops = -1, skip_indices = [1,] )

Make the output of get_poised_set even better suited for interpolation.

Input:

basis: A vector of polynomials constituting a Lagrange basis for the polynomial space.

points: The vector of points belonging to the Lagrange basis.

LAMBDA :: Real > 1: Determines the quality of the interpolation.

solver: NLopt solver to use. Should be derivative-free.

max_solver_evals: Maximum number of evaluations in each optimization run.

max_loops: Maximum number of loops that try to make the set Λ-poised.

skip_indices: Inidices of points to discard last.

Return:

poised_points :: Vector{T} where T is either a Vector{F} or an SVector{n_vars, F} and F is the precision of the points in points, but at least Float32.

lagrange_basis :: Vector{<:AbstractPolynomialLike}: The Lagrange basis corresponding to poised_points.

point_indices: An array indicating which points from points are also in poised_points. A positive entry corresponds to the index of a poised point in points. If a poised point is new, then the entry is -1.

Morbit.max_evals — Method

(Soft) upper bound on the number of function calls.

Morbit.new_result! — Function

Add a new result to the database, return its id of type Int.

Morbit.num_evals! — Method

Set evaluation counter to N.

Morbit.num_evals — Method

Number of calls to the original objective function.

Morbit.num_objectives — Method

Number of scalar-valued objectives of the problem.

Morbit.pop_objf! — Method

Remove objf from list_of_objectives(mop) and return its output indices.

Morbit.reset_evals! — Method

Set evaluation counter to 0 for each VectorObjectiveFunction in m.vector_of_objectives.

Morbit.results_in_box_indices — Function

Return indices of results in db that lie in a box with corners lb and ub.

Morbit.reverse_internal_sorting — Method

Sort an interal objective vector so that the objectives are in the order in which they were added.

Morbit.reverse_internal_sorting_indices — Method

Return index vector so that an internal objective vector is sorted according to the order the objectives where added.

Morbit.scale — Method

Scale variables fully constrained to a closed interval to [0,1] internally.

Morbit.set_transformed! — Method

Set the flag indicating whether the database data has been transformed or not.

Morbit.shrink_radius — Method

Shrink radius according to γ * Δ.

Morbit.shrink_radius — Method

Shrink radius according to γ * ||s||.

Morbit.shrink_radius_much — Method

Shrink radius much according to γ * Δ.

Morbit.shrink_radius_much — Method

Shrink radius according to γ * ||s||.

Morbit.unique_with_indices — Method

Return unique_elems, indices = unique_with_indices(arr) such that unique_elems[indices] == arr (and unique_elems == unique(arr)).

Morbit.unscale — Method

Reverse scaling for fully constrained variables from [0,1] to their former domain.