Developer Manual

This page is a big TODO!

For now, there are only the doc-strings:

Morbit.AbstractConfig — Type

Abstract super type for user configurable algorithm configuration.

Morbit.AbstractDB — Type

Abstract database super type. Implemented by ArrayDB and MockDB.

Morbit.AbstractDescentConfig — Type

Abstract super type for descent step configuration.

Morbit.AbstractIterSaveable — Type

Abstract super type for some saveable iteration information.

Morbit.AbstractIterate — Type

Abstract super type for a container that stores the site and value vectors.

Morbit.AbstractMOP — Type

AbstractMOP{T}

Abstract super type for multi-objective optimization problems. T is true if the problem is modifyable and `false elsewise.

The user should define a MOP<:AbstractMOP{true}, see MOP.

Morbit.AbstractSurrogate — Type

Abstract super type for the actual surrogate models.

Morbit.AbstractSurrogateConfig — Type

Abstract super type for a configuration type defining some surrogate model.

Morbit.AbstractSurrogateMeta — Type

Abstract super type for meta data that is used to build a model.

Morbit.AbstractVecFun — Type

Abstract super type for any kind of (vector) objective.

Morbit.CompositeVecFun — Type

CompositeVecFun(; inner_ref, outer_ref,
    inner_nl_index = nothing, outer_nl_index = nothing,
    num_outputs = num_outputs( outer_ref ) 
)

A CompositeVecFun is a special AbstractVecFun to use for functions $f:ℝⁿ→ℝᵏ, f = φ ∘ g$, where $g:ℝⁿ → ℝᵐ$ is an expensive function (that is possibly used elsewhere too) and $φ:ℝᵐ→ℝᵏ$ is an outer function. Hence, when evaluating a CompositeVecFun, the value of the function referenced by inner_ref is plugged into outer_ref. When evaluating the overall jacobian at $x$, the chain rule provides

\[\begin{aligned} ℝ^{k×n} \ni Df(x) &= Dφ(g(x))Dg(x), \quad Dφ ∈ ℝ^{k×m}, Dg ∈ ℝ^{m×n}, ∇f_ℓ(x) &= (Dφ(g(x))Dg(x))_{ℓ,:}^T = (∇φ_ℓ(g(x))^T Dg(x))^T = Dg(x)^T ∇φ_ℓ(g(x)).\]

For the Hessian of output $f_ℓ$ it holds that

\[Hf_ℓ(x) = D(∇f_ℓ(x)) = D(Dφ_ℓ(g(x)) \cdot Dg(x)) = (Hφ_ℓ(g(x)) \ Dg(x))\cdot Dg(x) + Dφ_ℓ(g(x)) \cdot Hg(x)\]

Morbit.ExactConfig — Type

ExactConfig(; gradients, jacobian = nothing, max_evals = typemax(Int64))

Configuration for an ExactModel. gradients should be a vector of callbacks for the objective gradients or a Symbol, either :autodiff or fdm, to define the differentiation method to use on the objective. Alternatively, a jacobian handle can be provided.

Morbit.ExactModel — Type

ExactModel( index, mop )

Exact Model type for evaluating the objective function objf directly. Is instantiated by the corresponding init_model and update_model functions.

Morbit.LagrangeConfig — Type

LagrangeConfig(; kwargs... )

Configuration for Lagrange Polyoniaml models.

degree

Degree of the surrogate model polynomials. Default: 2

θ_enlarge

Enlargement parameter to consider more points for inclusion. Default: 2.0

LAMBDA

Quality parameter in Λ-Poisedness Algorithm. Default: 1.5

allow_not_linear

Whether or not the interpolation sets must be Λ-poised (and the models fully linear). Default: false

optimized_sampling

Whether or not to try to construct a new interpolation set in each iteration. Default: true

save_path

Path to look at for a pre-saved poised set of interpolation points in [0,1]^n. Default:

io_lock

Lock to use parallel setups if multiple models can access save_path. Default: nothing

algo1_max_evals

Maximum number of polynomial evaluations to find a poised set. Default: -1

algo2_max_evals

Maximum number of polynomial evaluations to make a set Λ-poised. Default: -1

algo1_solver

NLopt Solver to use for the poisedness problem. Default: :LN_BOBYQA

algo2_solver

NLopt Solver to use for the Λ-poisedness problem. Default: :LN_BOBYQA

max_evals

Maximum number of evaluations allowed to the true objective(s). Default: typemax(Int64)

Morbit.RbfConfig — Type

RbfConfig(; kwarg1 = val1, … )

Configuration type for local RBF surrogate models.

kernel

RBF kernel (Symbol), either :cubic, :inv_multiquadric, :multiquadric, :exp or :thin_plate_spline. Default: :cubic

shape_parameter

RBF shape paremeter, either a number or a string containing Δ. Default: NaN

polynomial_degree

Degree of polynomial attached to RBF. -1 means no polynomial. Default: 1

θ_enlarge_1

Local enlargment factor of trust region for sampling. Default: 2

θ_enlarge_2

Maximum enlargment factor of maximum trust region for sampling. Default: 2

θ_pivot

Sampling parameter to generate Λ-poised set. The higher, the more poised. Default: 1 / (2θenlarge1)

θ_pivot_cholesky

Parameter for 2nd sampling algorithm to ensure boundedness of Cholesky factors. Default: 1.0e-7

require_linear

Require models to be fully linear in each iteration. Default: true

max_model_points

Maximum number of training sites. -1 is reset to 2n+1. Default: -1

use_max_points

Sample new sites to always use the maximum number of points. Default: false

optimized_sampling

Whether or not to re-construct the training set in each iteration. Default: true

max_evals

Maximum number of objective evaluations. Default: typemax(Int64)

Morbit.RefSurrogate — Type

RefSurrogate(model_ref, output_indices)

A RefSurrogate holds a reference model_ref to some AbstractSurrogate object and delegates most of the evaluation to this inner model. RefSurrogates are meant to model a single (vector-valued) objective or constraint function. As an AbstractSurrogate might model more than one objetive or constraint, we have a field output_indices for filtering out the right output indices from the inner model output.

RefVecFun

Morbit.RefVecFun — Type

RefVecFun( inner_ref, nl_index = nothing )

A RefVecFun simply stores a reference to a VecFun object. Evaluation methods are delegated to this object. nl_index is an optional information field used by MOP.

VecFun, MOP

Morbit.SuperDB — Type

SuperDB{T,Vector{<IterSaveable}}

A SuperDB is nothing but a directory of sub-databases (sub_dbs) that store evaluation data for tuples of AnyIndex functions. There is also a field iter_data that stores iteration data in a vector for later inspection.

Morbit.TaylorCallbackConfig — Type

TaylorCallbackConfig(;degree=1,max_evals=typemax(Int64))

Configuration for a linear or quadratic Taylor model where there are callbacks provided for the gradients and – if applicable – the Hessians.

Morbit.TaylorConfig — Type

TaylorConfig(; degree, gradients :: RFD.CFDStamp, hessians :: RFD.CFDStamp, max_evals)

Configuration for a polynomial Taylor model using finite difference approximations of the derivatives. By default we have degree = 2 and gradients == hessians == RFD.CFDStamp(1,2), that is, a first order central difference scheme of accuracy order 3 is recursed to compute the Hessians and the gradients. In this case, the finite difference scheme is the same for both Hessians and gradients and we profit from caching intermediate results.

Morbit.VecFun — Type

VecFun(; n_out, model_config, func_handle, diff_wrapper = nothing)

Wrap the function func_handle to ensure vector output. A VecFun also provides a unified differentiation API.

Base.length — Method

Number of entries in database.

Morbit._add_eq_constraint! — Method

Add a linear equality constraint function to the model.

Morbit._add_ineq_constraint! — Method

Add a linear inequality constraint function to the model.

Morbit._add_nl_eq_constraint! — Method

Add a nonlinear equality constraint function to the model.

Morbit._add_nl_ineq_constraint! — Method

Add a nonlinear inequality constraint function to the model.

Morbit._add_objective! — Method

Add an objective function to the model.

Morbit._add_result! — Method

Add result res to database db.

Morbit._backtrack — Method

Perform a backtracking loop starting at x with an initial step of step_size .* dir and return trial point x₊, the surrogate value-vector m_x₊ and the final step s = x₊ .- x.

Morbit._consume_points — Method

_consume_points(data_base, poised_points, poised_indices, candidate_indices)

Helper to return array of database indices for poised_points and poised_indices. Add result to database if index is -1. candidate_indices are the database indices of the points from the trust region.

Morbit._del! — Method

Remove a function from the MOP.

Morbit._equal_vals — Method

_equal_vals( r1 :: Result, r2 :: Result )

Return true if both the site and the value vectors of r1 and r2 are equal.

Morbit._eval_models — Method

Evaluate (internal) output ℓ of TaylorModel tm, provided a difference vector h = x - x0.

Morbit._get_global_dir — Method

_get_global_dir( cfg, fx )

Return the objective space direction r ≥ 0. If a direction is stored in cfg, then it is returned. If a global ideal point i is stored in cfg, then fx .- i is returned. Else nothing is returned.

Morbit._get_optim_handle — Method

Return a function handle to be used with NLopt for output ℓ of model. That is, if model is a surrogate for two scalar objectives, then ℓ must be either 1 or 2.

Morbit._get_ps_objective_func — Method

Return objective function for Pascoletti-Serafini, modifying gradient in place.

Morbit._init_iterate — Method

_init_iterate( T , x, x_scaled, fx, 
    l_e, l_i, c_e, c_i, Δ, x_index_mapping )

Return an AbstractIterate of type T. The arguments x through c_i are site and value vectors. Δ is the associated trust region radius. x_index_mapping is a Dict mapping FunctionIndexTuples to the index of x in the current (sub)database.

Morbit._intersect_bounds — Method

Return the maximum or minimum stepsize $σ$ such that $x + σd$ conforms to the linear constraints $lb ≤ x+σd ≤ ub$ and $A(x+σd) - b ≦ 0$.

Morbit._local_bounds — Method

Return lower and upper bound vectors combining global and trust region constraints.

Morbit._missing_ids — Method

Return vector of ids of database db that are not evaluated yet.

Morbit._set_delta! — Method

Set current trust region radius (vector?) to Δ.

Morbit._surrogate_from_vec_function — Method

_surrogate_from_vec_function( vfun, scal, gs, nl_ind )

Provided a vector function vfun that is either of type RefVecFun or CompositeVecFun and a GroupedSurrogates object gs, that contains the model for vfun.nl_index, create and return a RefSurrogate or CompositeSurrogate from gs.

Morbit._transform_linear_constraints — Method

Return the linear constraint matrices $Aₛ, bₛ$ for the scaled domain, i.e., matrices such that $Aₛ xₛ ≤ bₛ$.

Morbit._typename — Function

Return the basis typename, i.e., remove all parameters from a typename.

Morbit.add_eq_constraint! — Function

add_eq_constraint!(mop, A, b = [])

Add linear equality constraints $Ax = b$ to the problem mop. If b is empty, then $Ax = 0$ is added. Returns a ConstraintIndex.

Morbit.add_ineq_constraint! — Function

add_ineq_constraint!(mop, A, b = [])

Add linear inequality constraints $Ax ≤ b$ to the problem mop. If b is empty, then $Ax ≤ 0$ is added. Returns a ConstraintIndex.

Morbit.do_groupings — Method

Group functions with indices of type NLIndex by model config type.

Morbit.ensure_contains_res_with_site! — Method

Return id of result in db with site x and values y. Create if necessary.

Morbit.ensure_contains_values! — Method

Return id of result in db with site x and values y. Create if necessary.

Morbit.eval_missing! — Method

Evaluate all unevaluated results in db using objectives of mop.

Morbit.eval_missing! — Method

Evaluate all unevaluated results in db using objectives of mop.

Morbit.eval_models — Method

eval_models( mod, scal, x_scaled, ℓ)

Evaluate output(s) ℓ of model mod at scaled site x_scaled.

Morbit.eval_models — Method

Evaluate the ExactModel em at scaled site x̂.

Morbit.eval_models — Method

Evaluate (internal) output(s) ℓ of tm at scaled site x̂.

Morbit.eval_models — Method

Evaluate output ℓ of mod::RbfModel at scaled site x̂.

Morbit.eval_models — Method

Evaluate mod::RbfModel at scaled site x̂.

Morbit.find_result — Function

find_result(db, x, y)

Return id of a result in db that has site x and value y or return -1 if there is no such result.

Morbit.full_lower_bounds — Method

Return full vector of lower variable vectors for original problem.

Morbit.full_upper_bounds — Method

Return full vector of upper variable vectors for original problem.

Morbit.get_delta — Method

Return current trust region radius (vector) $Δᵗ$.

Morbit.get_eq_const — Method

Return value vector of linear equality constraints.

Morbit.get_eq_constraint_indices — Method

Return the vector or tuple of the indices of linear equality constraints used in the model.

Morbit.get_evaluated_flag — Method

Return true if the result with id in db has a valid evaluation vector.

Morbit.get_fx — Method

Return current value vector $f(xᵗ)$.

Morbit.get_gradient — Method

Gradient vector of output ℓ of em at scaled site x̂.

Morbit.get_gradient — Method

Gradient vector of output ℓ of mod at scaled site x̂.

Morbit.get_id — Method

get_id( res :: Result ) :: Int

Return the id of a result such that for the database db containing res it holds that get_result(db, id) == res.

Morbit.get_ids — Method

List of all id :: Int belonging to the stored results.

Morbit.get_ineq_const — Method

Return value vector of linear inequality constraints.

Morbit.get_ineq_constraint_indices — Method

Return the vector or tuple of the indices of linear inequality constraints used in the model.

Morbit.get_jacobian — Function

Jacobian Matrix of ExactModel em at scaled site x̂.

Morbit.get_lower_bound — Method

Return the lower bound for the variable with Index VarInd.

Morbit.get_model_jacobian — Function

Jacobian Matrix of ExactModel em at scaled site x̂.

Morbit.get_nl_eq_const — Method

Return current equality constraint vector $cₑ(xᵗ)$.

Morbit.get_nl_eq_constraint_indices — Method

Return the vector or tuple of the indices of nonlinear equality constraints used in the model.

Morbit.get_nl_ineq_const — Method

Return current inequality constraint vector $cᵢ(xᵗ)$.

Morbit.get_nl_ineq_constraint_indices — Method

Return the vector or tuple of the indices of nonlinear inequality constraints used in the model.

Morbit.get_objective_indices — Method

Return a vector or tuple of all objective indices used in the model.

Morbit.get_poised_set — Method

get_poised_set( basis, points; solver = :LN_BOBYQA, max_solver_evals = -1 )

Compute a point set suited for polynomial interpolation.

Input:

basis: A vector of polynomials constituting a basis for the polynomial space.

points: (optional) A set of candidate points to be tried for inclusion into the poised set.

solver: NLopt solver to use. Should be derivative-free.

max_solver_evals: Maximum number of evaluations in each optimization run.

Return:

poised_points :: Vector{T} where T is either a Vector{F} or an SVector{n_vars, F} and F is the precision of the points in points, but at least Float32.

lagrange_basis :: Vector{<:AbstractPolynomialLike}: The Lagrange basis corresponding to poised_points.

point_indices: An array indicating which points from points are also in poised_points. A positive entry corresponds to the index of a poised point in points. If a poised point is new, then the entry is -1.

Morbit.get_res_type — Method

Return type of results stored in database.

Morbit.get_result — Method

Get result with id from database db.

Morbit.get_saveable_type — Method

Return type of AbstractIterSaveables stored in database.

Morbit.get_scaling_poly — Method

Return vector of polynomials that scales variables from [lb, ub] to [0,1]^n.

Morbit.get_site — Method

Return the evaluation value vector for result with id in database db.

Morbit.get_site — Method

get_site( res :: Result{XT,YT} )

Return evaluation site of type XT associated with res.

Morbit.get_sites — Method

Return a vector of all evaluation site vectors stored in database.

Morbit.get_unscaling_poly — Method

Return vector of polynomials that unscales variables from [0,1]^n to [lb,ub].

Morbit.get_upper_bound — Method

Return the upper bound for the variable with Index VarInd.

Morbit.get_vals — Method

Return only the parts of f(x) that are relevant for func_indices.

Morbit.get_value — Method

Return the evaluation site vector for result with id in database db.

Morbit.get_value — Method

get_value( res :: Result{XT,YT} )

Return evaluation value vector of type YT associated with res.

Morbit.get_values — Method

Return a vector of all evaluation value vectors stored in database.

Morbit.get_x — Method

Return current iteration site vector $xᵗ$.

Morbit.get_x_index — Method

Index (or id) of current iterate in database.

Morbit.get_x_scaled — Method

Return current iteration site vector $xᵗ$.

Morbit.grow_radius — Method

Grow radius according to min( Δ_max, γ * Δ ).

Morbit.grow_radius — Method

Grow radius according to min( Δ_max, (γ + ||s||/Δ) * Δ )

Morbit.has_valid_site — Method

has_valid_site( r :: Result )

Return true if the site vector of r is neither empty nor NaN.

Morbit.has_valid_value — Method

has_valid_value( r :: Result )

Return true if the value vector of r is neither empty nor NaN.

Morbit.init_db — Method

Constructor for empty database of type MockDB.

Morbit.init_db — Method

Constructor for empty database of type ArrayDB.

Morbit.init_iterate — Method

init_iter_data( T , x, fx, Δ )

Return an instance of "base" type T implementing AbstractIterate with correct type parameters for x, fx and Δ. x and fx should be vectors of floats and Δ can either be a float or a vector of floats.

Morbit.init_model — Method

Return an ExactModel build from a VecFun objf. Model is the same inside and outside of criticality round.

Morbit.init_res — Function

init_res( id, x, y)

Return of result with site x, value y and database id id.

Morbit.init_surrogates — Method

Return a SurrogateContainer initialized from the information provided in mop.

Morbit.is_transformed — Method

Bool indicating if the database data been transformed.

Morbit.local_bounds — Method

Local bounds vectors lb_eff and ub_eff using scaled variable constraints from mop.

Morbit.make_outer_fun — Method

make_outer_fun( expr_str :: String; n_vars, kwargs...)

Helper function to build a VecFun from a an expression string.

Morbit.make_outer_fun — Method

make_outer_fun( func :: Function; n_vars, n_out, kwargs...)

Helper function to build a VecFun from a function func taking two vector valued arguments: First, the argument vector $x$ of length n_vars and second, the value of an inner function evaluated at $x$.

Morbit.make_set_lambda_poised — Method

make_set_lambda_poised( basis, points; 
    LAMBDA = 1.5, solver = :LN_BOBYQA, max_solver_evals = -1, max_loops = -1, skip_indices = [1,] )

Make the output of get_poised_set even better suited for interpolation.

Input:

basis: A vector of polynomials constituting a Lagrange basis for the polynomial space.

points: The vector of points belonging to the Lagrange basis.

LAMBDA :: Real > 1: Determines the quality of the interpolation.

solver: NLopt solver to use. Should be derivative-free.

max_solver_evals: Maximum number of evaluations in each optimization run.

max_loops: Maximum number of loops that try to make the set Λ-poised.

skip_indices: Inidices of points to discard last.

Return:

poised_points :: Vector{T} where T is either a Vector{F} or an SVector{n_vars, F} and F is the precision of the points in points, but at least Float32.

lagrange_basis :: Vector{<:AbstractPolynomialLike}: The Lagrange basis corresponding to poised_points.

point_indices: An array indicating which points from points are also in poised_points. A positive entry corresponds to the index of a poised point in points. If a poised point is new, then the entry is -1.

Morbit.make_vec_fun — Method

make_vec_fun( fn; 
model_cfg, n_out, can_batch = false,
gradients = nothing, jacobian = nothing, hessian = nothing,
diff_method = FiniteDiffWrapper )

Pack the function fn::Function into a VecFun and ensure that appropriate derivative information can be queried, as needed by model_cfg.

Morbit.make_vec_fun — Method

make_vec_fun( fn; 
model_cfg, n_out, can_batch = false,
gradients = nothing, jacobian = nothing, hessian = nothing,
diff_method = FiniteDiffWrapper )

Pack the function fn::CountedFunc into a VecFun and ensure that appropriate derivative information can be queried, as needed by model_cfg.

Morbit.max_evals — Method

(Soft) upper bound on the number of function calls.

Morbit.new_result! — Method

Add a new result to the database, return its id of type Int.

Morbit.next_id — Method

Return an id for the next result to be added to db.

Morbit.nullify_last_row — Method

nullify_last_row( R )

Returns matrices $B$ and $G$ with $G⋅R = B$. $G$ applies givens rotations to make $R$ upper triangular. $R$ is assumed to be an upper triangular matrix augmented by a single row.

Morbit.num_objectives — Method

Number of scalar-valued objectives of the problem.

Morbit.register_func — Method

register_func(func, func_name)

Registers the function func for subsequent use in a function expression. Note, that using function expressions is really only advisable if the "base function" is truly expensive and surrogate modelling remedies the performance penalties from parsing strings and @evaling expressions.

Morbit.reset_evals! — Method

Set evaluation counter to 0 for each VecFun in m.vector_of_objectives.

Morbit.results_in_box_indices — Function

Return indices of results in db that lie in a box with corners lb and ub.

Morbit.set_evaluated_flag! — Function

Set the evaluation status for result with id to state.

Morbit.set_site! — Method

Set site of result with id in database db to x.

Morbit.set_transformed! — Method

Set the flag indicating whether the database data has been transformed or not.

Morbit.set_value! — Method

Set value of result with id in database db to x.

Morbit.shrink_radius — Method

Shrink radius according to γ * Δ.

Morbit.shrink_radius — Method

Shrink radius according to γ * ||s||.

Morbit.shrink_radius_much — Method

Shrink radius much according to γ * Δ.

Morbit.shrink_radius_much — Method

Shrink radius according to γ * ||s||.

Morbit.stamp! — Method

stamp!(db, ids)

Put the saveable ids into the database db.

Morbit.transform! — Method

Scale the site of result with id in database db using bounds of mop.

Morbit.transform! — Method

Apply scaling and objectives sorting to each result in database db.

Morbit.unique_with_indices — Method

Return unique_elems, indices = unique_with_indices(arr) such that unique_elems[indices] == arr (and unique_elems == unique(arr)).

Morbit.untransform! — Method

Unscale the site of result with id in database db using bounds of mop.

Morbit.untransform! — Method

Undo scaling and objectives sorting to each result in database db.

Morbit.var_indices — Method

Return a vector of VariableIndices used in the model.